Base/OrderStatusDataFixuture removed
- all order statuses are created in database migration
- order statuses persistence references are created in new Demo/OrderStatusDataFixuture
