+++ packages/framework/src/DataFixtures/DemoMultidomain/OrderDataFixture.php
@@ -7,8 +7,8 @@ use Doctrine\Common\Persistence\ObjectManager;
 use Faker\Generator;
 use Shopsys\FrameworkBundle\Component\DataFixture\AbstractReferenceFixture;
 use Shopsys\FrameworkBundle\DataFixtures\Base\CurrencyDataFixture;
-use Shopsys\FrameworkBundle\DataFixtures\Base\OrderStatusDataFixture;
 use Shopsys\FrameworkBundle\DataFixtures\Demo\OrderDataFixture as DemoOrderDataFixture;
+use Shopsys\FrameworkBundle\DataFixtures\Demo\OrderStatusDataFixture;
 use Shopsys\FrameworkBundle\DataFixtures\Demo\PaymentDataFixture as DemoPaymentDataFixture;
 use Shopsys\FrameworkBundle\DataFixtures\Demo\ProductDataFixture as DemoProductDataFixture;
 use Shopsys\FrameworkBundle\DataFixtures\Demo\TransportDataFixture as DemoTransportDataFixture;
