+++ packages/framework/src/DataFixtures/Performance/OrderDataFixture.php
@@ -8,8 +8,8 @@ use Shopsys\FrameworkBundle\Component\Console\ProgressBarFactory;
 use Shopsys\FrameworkBundle\Component\DataFixture\PersistentReferenceFacade;
 use Shopsys\FrameworkBundle\Component\Doctrine\SqlLoggerFacade;
 use Shopsys\FrameworkBundle\DataFixtures\Base\CurrencyDataFixture;
-use Shopsys\FrameworkBundle\DataFixtures\Base\OrderStatusDataFixture;
 use Shopsys\FrameworkBundle\DataFixtures\Demo\CountryDataFixture;
+use Shopsys\FrameworkBundle\DataFixtures\Demo\OrderStatusDataFixture;
 use Shopsys\FrameworkBundle\DataFixtures\Demo\PaymentDataFixture;
 use Shopsys\FrameworkBundle\DataFixtures\Demo\TransportDataFixture;
 use Shopsys\FrameworkBundle\DataFixtures\Performance\ProductDataFixture as PerformanceProductDataFixture;
